+++ src/services/AuthService.ts
@@ -1,6 +1,10 @@
-import { OpaqueToken } from "angular2/core";
+import { OpaqueToken, EventEmitter } from "angular2/core";
 
 export interface Class {
+	loggedInEmitter:EventEmitter<any>;
+	loggedOutEmitter:EventEmitter<any>;
+	authChangedEmitter:EventEmitter<any>;
+
 	isAuthenticated():boolean;
 
 	login( username:string, password:string, rememberMe:boolean ):Promise<any>;